refactor: Update file paths for main.rs and builders
This commit updates the file paths for the main.rs file and the builders module. The main.rs file has been moved from src/main.rs to src/bin/main.rs. Additionally, the builders module has been moved from src/backlog_builder.rs, src/rust_builder.rs, and src/web_builder.rs to src/builders/backlog_builder.rs, src/builders/rust_builder.rs, and src/builders/web_builder.rs respectively. These changes were made to improve the organization and structure of the codebase.
