+++ elmio-rs/elmio-cli/src/lib.rs
@@ -1,13 +1,4 @@
-pub mod asset_hasher;
-pub mod backlog_builder;
-pub mod build;
-pub mod cleaner;
-pub mod exec;
-pub mod file_util;
+pub mod builders;
+pub mod commands;
 pub mod project;
-pub mod project_info;
-pub mod rust_builder;
-pub mod script_runner;
-pub mod serve;
-pub mod watch;
-pub mod web_builder;
+pub mod utils;
